fix(j-s): refactor component prop drilling and add animation
